Remove populus dependency
This removes the dependency of populus to build our contracts and uses
our new own tool to build the contracts.
