+++ .gitignore
@@ -1,5 +1,4 @@
 *~
-/venv-populus/
 /venv/
 .pytest_cache/
 # Byte-compiled / optimized / DLL files
