+++ .soliumignore
@@ -1,4 +1,3 @@
 node_modules
 build
-venv-populus
 venv
