+++ MANIFEST.in
@@ -1,4 +1,3 @@
 include README.rst
 include requirements.txt
 recursive-include contracts *.sol
-include populus.json
