--- constraints-populus.txt
@@ -1,55 +1,0 @@
-anyconfig==0.9.7
-argh==0.26.2
-atomicwrites==1.1.5
-attrs==18.1.0
-bitcoin==1.1.42
-certifi==2018.4.16
-cffi==1.11.5
-chardet==3.0.4
-click==6.7
-contextlib2==0.5.5
-cytoolz==0.9.0.1
-eth-abi==0.5.0
-eth-keys==0.1.0b4
-eth-tester==0.1.0b11
-eth-testrpc==1.3.5
-eth-utils==0.7.4
-ethereum==1.6.1
-flake8==3.5.0
-idna==2.7
-json-rpc==1.11.0
-jsonschema==2.6.0
-lru-dict==1.1.6
-mccabe==0.6.1
-more-itertools==4.2.0
-pathtools==0.1.2
-pbkdf2==1.3
-pluggy==0.6.0
-populus==2.2.0
-py==1.5.4
-py-geth==2.0.1
-py-solc==3.1.0
-pycodestyle==2.3.1
-pycparser==2.18
-pycrypto==2.6.1
-pycryptodome==3.6.4
-pyethash==0.1.27
-pyflakes==1.6.0
-pylru==1.1.0
-pysha3==1.0.2
-pytest==3.2.5
-PyYAML==3.13
-repoze.lru==0.7
-requests==2.19.1
-rlp==0.6.0
-scrypt==0.8.6
-secp256k1==0.13.2
-semantic-version==2.6.0
-six==1.11.0
-texttable==1.2.1
-toolz==0.9.0
-toposort==1.5
-urllib3==1.23
-watchdog==0.8.3
-web3==3.16.5
-Werkzeug==0.14.1
