+++ constraints.txt
@@ -1,4 +1,5 @@
 aniso8601==3.0.2
+apipkg==1.5
 appdirs==1.4.3
 asn1crypto==0.24.0
 atomicwrites==1.2.1
@@ -13,6 +14,7 @@ certifi==2018.8.24
 cffi==1.11.5
 chardet==3.0.4
 click==7.0
+contract-deploy-tools==0.1.1
 coverage==4.5.1
 cryptography==2.3.1
 cytoolz==0.9.0.1
@@ -27,6 +29,7 @@ eth-rlp==0.1.2
 eth-tester==0.1.0b32
 eth-typing==1.3.0
 eth-utils==1.2.1
+execnet==1.5.0
 firebase-admin==2.13.0
 flake8==3.5.0
 flake8-invalid-escape-sequences==1.3
@@ -78,6 +81,7 @@ ptyprocess==0.6.0
 py==1.6.0
 py-ecc==1.4.3
 py-evm==0.2.0a32
+py-solc==3.2.0
 pyasn1==0.4.4
 pyasn1-modules==0.2.2
 pycodestyle==2.3.1
