+++ requirements.txt
@@ -1,11 +1,12 @@
 web3
 click
-# --- development dependencies, i.e. dependencies not needed for running ethindex
+# --- development dependencies, i.e. dependencies not needed for running contracts
 flake8
 pytest
 texttable
 setuptools_scm
 eth-tester[py-evm]
+contract-deploy-tools
 # require a recent pip version, otherwise make install may silently fail to
 # install the newly compiled contracts
 pip>=18.1
