+++ tests/conftest.py
@@ -1,9 +1,66 @@
+import os
+import subprocess
+import shutil
+import io
+from pathlib import Path
+
 import pytest
 from web3 import Web3
 from web3.providers.eth_tester import EthereumTesterProvider
 from eth_utils import to_checksum_address, encode_hex
 import eth_tester
 import eth_tester.backends.pyevm.main
+
+
+@pytest.fixture(scope="session", autouse=True)
+def compile_contracts():
+    """compile the contracts before running the tests"""
+    cwd = os.getcwd()
+    try:
+        root = Path(__file__).parent.parent
+        os.chdir(root)
+        err = os.system("deploy-tools compile -o")
+        assert err == 0, "Compilation failed"
+        os.environ["TRUSTLINES_CONTRACTS_JSON"] = str(root / "build" / "contracts.json")
+    finally:
+        os.chdir(cwd)
+
+
+def _find_solc(msgs):
+    solc = shutil.which("solc")
+    if solc:
+        msgs.write("solc: {}\n".format(solc))
+    else:
+        msgs.write("solc: <NOT FOUND>\n")
+    return solc
+
+
+def _get_solc_version(msgs):
+    try:
+        process = subprocess.Popen(['solc', '--version'], stdout=subprocess.PIPE)
+    except Exception as err:
+        msgs.write("solidity version: <ERROR {}>".format(err))
+        return
+
+    out, err = process.communicate()
+
+    lines = out.decode('utf-8').splitlines()
+    for line in lines:
+        if line.startswith("Version: "):
+            msgs.write("solidity version: {}\n".format(line[len("Version: "):]))
+            break
+    else:
+        msgs.write("solidity version: <UNKNOWN>")
+
+
+def pytest_report_header(config):
+    msgs = io.StringIO()
+    solc = _find_solc(msgs)
+
+    if solc:
+        _get_solc_version(msgs)
+
+    return msgs.getvalue()
 
 
 @pytest.fixture(scope="session", autouse=True)
